Further generalise cascade and move it to PromiseNotifier
