+++ transport/src/main/java/io/netty/channel/ChannelPromiseNotifier.java
@@ -15,14 +15,7 @@
  */
 package io.netty.channel;
 
-import io.netty.util.concurrent.Future;
-import io.netty.util.concurrent.FutureListener;
-import io.netty.util.concurrent.GenericFutureListener;
-import io.netty.util.concurrent.Promise;
 import io.netty.util.concurrent.PromiseNotifier;
-import io.netty.util.internal.PromiseNotificationUtil;
-import io.netty.util.internal.logging.InternalLogger;
-import io.netty.util.internal.logging.InternalLoggerFactory;
 
 /**
  * ChannelFutureListener implementation which takes other {@link ChannelPromise}(s) and notifies them on completion.
@@ -52,63 +45,5 @@ public final class ChannelPromiseNotifier
      */
     public ChannelPromiseNotifier(boolean logNotifyFailure, ChannelPromise... promises) {
         super(logNotifyFailure, promises);
-    }
-
-    /**
-     * Link the {@link ChannelFuture} and {@link Promise} such that if the {@link ChannelFuture} completes the {@link
-     * Promise} will be notified, and completed with the channel in question. Cancellation is propagated both ways such
-     * that if the {@link Future} is cancelled the {@link Promise} is cancelled and vise-versa.
-     *
-     * @param future  the {@link ChannelFuture} which will be used to listen to for notifying the {@link Promise}.
-     * @param promise the {@link Promise} which will be notified
-     * @return the passed in {@link ChannelFuture}
-     */
-    public static ChannelFuture cascadeChannel(ChannelFuture future, final Promise<Void> promise) {
-        return cascadeChannel(true, future, promise, null);
-    }
-
-    /**
-     * Link the {@link Future} and {@link Promise} such that if the {@link Future} completes the {@link Promise} will be
-     * notified. Cancellation is propagated both ways such that if the {@link Future} is cancelled the {@link Promise}
-     * is cancelled and vise-versa.
-     *
-     * @param logNotifyFailure {@code true} if logging should be done in case notification fails.
-     * @param future           the {@link Future} which will be used to listen to for notifying the {@link Promise}.
-     * @param promise          the {@link Promise} which will be notified
-     * @param successResult    the result that will be propagated to the promise on success
-     * @return the passed in {@link Future}
-     */
-    public static <R> ChannelFuture cascadeChannel(boolean logNotifyFailure, ChannelFuture future,
-                                                   Promise<R> promise, R successResult) {
-        promise.addListener(new FutureListener<Object>() {
-            @Override
-            public void operationComplete(Future<Object> f) {
-                if (f.isCancelled()) {
-                    future.cancel(false);
-                }
-            }
-        });
-        future.addListener(new GenericFutureListener<ChannelFuture>() {
-            @Override
-            public void operationComplete(ChannelFuture f) throws Exception {
-                if (promise.isCancelled() && f.isCancelled()) {
-                    // Just return if we propagate a cancel from the promise to the future and both are notified already
-                    return;
-                }
-                if (f.isSuccess()) {
-                    promise.setSuccess(successResult);
-                } else if (f.isCancelled()) {
-                    InternalLogger internalLogger = null;
-                    if (logNotifyFailure) {
-                        internalLogger = InternalLoggerFactory.getInstance(PromiseNotifier.class);
-                    }
-                    PromiseNotificationUtil.tryCancel(promise, internalLogger);
-                } else {
-                    Throwable cause = future.cause();
-                    promise.tryFailure(cause);
-                }
-            }
-        });
-        return future;
     }
 }
